// src/lib/integrations/platforms/github/webhook-schemas.ts
export const PushEventPayloadSchema = z.object({
  ref: z.string(),
  repository: z.object({
    full_name: z.string(),
  }),
  deleted: z.boolean(),
});
// ...
export type PushEventPayload = z.infer<typeof PushEventPayloadSchema>;